Refactoring: move all flag/command verification code to its own package, so now docker.go is almost only about the main function.
